// fe-app-podkop/src/podkop/types.ts
export namespace Podkop {
// ...
  export enum AvailableMethods {
    CHECK_DNS_AVAILABLE = 'check_dns_available',
    CHECK_FAKEIP = 'check_fakeip',
    CHECK_NFT_RULES = 'check_nft_rules',
    CHECK_ZAPRET_RUNTIME = 'check_zapret_runtime',
    CHECK_BYEDPI_RUNTIME = 'check_byedpi_runtime',
    CHECK_INBOUNDS_CONFIG = 'check_inbounds_config',
    GET_STATUS = 'get_status',
    GET_OUTBOUND_LINK = 'get_outbound_link',
    GET_OUTBOUND_LINK_STATES = 'get_outbound_link_states',
    GET_OUTBOUND_METADATA = 'get_outbound_metadata',
    GET_SUBSCRIPTION_METADATA = 'get_subscription_metadata',
    CHECK_SING_BOX = 'check_sing_box',
    CHECK_INBOUNDS = 'check_inbounds',
    GET_SING_BOX_STATUS = 'get_sing_box_status',
    GET_ZAPRET_STATUS = 'get_zapret_status',
    GET_BYEDPI_STATUS = 'get_byedpi_status',
    CLASH_API = 'clash_api',
    RESTART = 'restart',
    START = 'start',
    STOP = 'stop',
    ENABLE = 'enable',
    DISABLE = 'disable',
    GLOBAL_CHECK = 'global_check',
    SHOW_SING_BOX_CONFIG = 'show_sing_box_config',
    CHECK_LOGS = 'check_logs',
    CHECK_SING_BOX_LOGS = 'check_sing_box_logs',
    GET_SYSTEM_INFO = 'get_system_info',
    COMPONENT_ACTION = 'component_action',
    COMPONENT_ACTION_ASYNC = 'component_action_async',
    COMPONENT_ACTION_STATUS = 'component_action_status',
    SUBSCRIPTION_UPDATE = 'subscription_update',
    SUBSCRIPTION_UPDATE_ASYNC = 'subscription_update_async',
    SUBSCRIPTION_UPDATE_STATUS = 'subscription_update_status',
    BACKUP_CONFIG = 'backup_config',
    HUB_LIST_SOURCES = 'hub_list_sources',
    HUB_GET_MODULES = 'hub_get_modules',
  }
// ...
  export interface MethodSuccessResponse<T> {
    success: true;
    data: T;
  }

  export interface MethodFailureResponse {
    success: false;
    error: string;
  }

  export type MethodResponse<T> =
    | MethodSuccessResponse<T>
    | MethodFailureResponse;
// ...
}